examples/mutual_guilds/requirements.txt
hikari==2.0.0.dev104
quart==0.16.2
python-dotenv==0.19.2